+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## NEXT RELEASE
 
+* Added the `author` to the messages
+* Switched from `assignees` to `requested reviewers` for the `waiting on` portion of the message
 * Added `base_url` paremeter to specify an enterprise GitHub instance if necessary
 * Reworked the `github_token` logic to allow for better unauthenticated usage when no GitHub token is passed
+* Corrects the `Messages` class name to the singular `Message`
 
 ## v3.2.1 (2022-01-26)